# SpringEcom/src/main/java/com/sharma/SpringEcom/model/dto/OrderResponse.java
package com.sharma.SpringEcom.model.dto;

import java.time.LocalDate;
import java.util.List;

public record OrderResponse(
        String orderId,
        String customerName,
        String email,
        String status,
        LocalDate orderDate,
        List<OrderItemResponse> items
) {
}